feat: support router plugin in rspack
