+++ .changeset/tough-queens-repeat.md
@@ -1,0 +1,6 @@
+---
+'@modern-js/app-tools': minor
+---
+
+feat: support router-plugin in rspack bundler.
+feat: 在 rspack bundler 里支持 router-plugin
