--- packages/solutions/app-tools/src/builder/builder-webpack/webpackPlugins/index.ts
@@ -1,1 +1,0 @@
-export * from './RouterPlugin';
