+++ packages/solutions/app-tools/src/builder/shared/bundlerPlugins/index.ts
@@ -1,0 +1,3 @@
+export * from './HtmlAsyncChunkPlugin';
+export * from './HtmlBottomTemplate';
+export * from './RouterPlugin';
